requirements.txt
flask==2.0.1
flask-cors==3.0.10
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pdfkit==1.0.0
spacy==3.1.3
scikit-learn==1.0
pydantic==1.8.2
python-dateutil==2.8.2
PyPDF2==1.26.0